Changed variable naming to unsigned

+++ src/gravity/step.h
@@ -3,7 +3,7 @@
 #include <vector>
 
 inline void Step(std::vector<Particle>& p, real dt) {
-    const int count = p.size();
+    const size_t count = p.size();
     for (size_t i = 0; i < count; i++)
     {
         for (size_t j = i + 1; j < count; j++)
